Create arduino-connector.ino
This file contains the code for the Arduino to grab the data and send it to a serial which will be collected by CoolTerm software and saved to a data file. A python file later may access this file and send it to the thingSpeak channel with the API-key and channel id.